+++ libs/commands/publish/src/command.ts
@@ -1,4 +1,4 @@
-import log from "npmlog";
+import { log } from "@lerna/core";
 import type { CommandModule } from "yargs";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
@@ -106,6 +106,18 @@ const command: CommandModule = {
       "temp-tag": {
         describe: "Create a temporary tag while publishing.",
         type: "boolean",
+      },
+      throttle: {
+        describe: "Throttle module publication. This is implicit if a throttle size or delay is provided",
+        type: "boolean",
+      },
+      "throttle-size": {
+        describe: "Bucket size used to throttle module publication.",
+        type: "number",
+      },
+      "throttle-delay": {
+        describe: "Delay between throttle bucket items publications (in seconds).",
+        type: "number",
       },
       "no-verify-access": {
         // proxy for --verify-access