+++ libs/core/src/lib/command/index.ts
@@ -1,9 +1,11 @@
 import { ProjectFileMap } from "@nx/devkit";
+import { ExecOptions as NodeExecOptions } from "child_process";
 import cloneDeep from "clone-deep";
 import dedent from "dedent";
-import log from "npmlog";
 import { daemonClient } from "nx/src/daemon/client/client";
 import os from "os";
+import yargs from "yargs";
+import log, { Logger } from "../npmlog";
 import { CommandConfigOptions, Project } from "../project";
 import { ProjectGraphWithPackages } from "../project-graph-with-packages";
 import { ValidationError } from "../validation-error";
@@ -14,8 +16,6 @@ import { detectProjects } from "./detect-projects";
 import { isGitInitialized } from "./is-git-initialized";
 import { logPackageError } from "./log-package-error";
 import { warnIfHanging } from "./warn-if-hanging";
-import yargs from "yargs";
-import { ExecOptions as NodeExecOptions } from "child_process";
 
 /**
  * Execa compatible options type
@@ -32,7 +32,7 @@ const DEFAULT_CONCURRENCY = os.cpus().length;
 /**
  * Specific logger with log-level success enabled in order to use function without index signature
  */
-export interface LernaLogger extends log.Logger {
+export interface LernaLogger extends Logger {
   /**
    * Log with level success
    * @param prefix