+++ libs/core/src/lib/npmlog/gauge/set-interval.ts
@@ -1,0 +1,10 @@
+/* eslint-disable */
+// @ts-nocheck
+
+/**
+ * Inlined from deprecated package https://github.com/npm/gauge/blob/f8092518a47ac6a96027ae3ad97d0251ffe7643b
+ */
+
+"use strict";
+// this exists so we can replace it during testing
+module.exports = setInterval;